Extract getCellSize
